fix(initialize): add delay before loading config
Prior to this commit, within the `initialize()` function, the configuration
structure may not have been completely written to the EEPROM before it was
loaded. This could manifest, e.g., as a newly specified I2C address not
persisting.

In this commit, add a 10 millisecond delay after writing the configuration
structure to EEPROM before loading it. In practice, this seems to resolve the
issue.